PC-29804 feat(HomeHeader): split Banner from Header
