+++ src/features/home/components/headers/HomeHeader.native.test.tsx
@@ -71,11 +71,9 @@ describe('HomeHeader', () => {
         refetchUser: jest.fn(),
       }
       mockUseAuthContext.mockReturnValueOnce(useAuthContextResultMock)
-      mockUseAuthContext.mockReturnValueOnce(useAuthContextResultMock)
-      mockUseAuthContext.mockReturnValueOnce(useAuthContextResultMock)
 
       mockUseAvailableCredit.mockReturnValueOnce(credit)
-      mockUseAvailableCredit.mockReturnValueOnce(credit)
+
       mockGeolocBannerFromBackend()
 
       renderHomeHeader()
@@ -84,69 +82,6 @@ describe('HomeHeader', () => {
       expect(await screen.findByText(subtitle)).toBeOnTheScreen()
     }
   )
-
-  it('should display SignupBanner when user is not logged in', async () => {
-    const useAuthContextNotLoggedInMock = {
-      isLoggedIn: false,
-      isUserLoading: false,
-      setIsLoggedIn: jest.fn(),
-      refetchUser: jest.fn(),
-    }
-    mockUseAuthContext.mockReturnValueOnce(useAuthContextNotLoggedInMock)
-    mockGeolocBannerFromBackend()
-
-    renderHomeHeader()
-    await act(async () => {})
-
-    expect(await screen.findByText('Débloque ton crédit')).toBeOnTheScreen()
-  })
-
-  it('should display activation banner with BicolorUnlock icon when banner api call return activation banner', async () => {
-    mockBannerFromBackend({
-      banner: {
-        name: BannerName.activation_banner,
-        text: 'à dépenser sur l’application',
-        title: 'Débloque tes 1000\u00a0€',
-      },
-    })
-
-    renderHomeHeader()
-
-    expect(await screen.findByText('Débloque tes 1000\u00a0€')).toBeOnTheScreen()
-    expect(screen.getByText('à dépenser sur l’application')).toBeOnTheScreen()
-    expect(screen.getByTestId('BicolorUnlock')).toBeOnTheScreen()
-  })
-
-  it('should display activation banner with ArrowAgain icon when banner api call return retry_identity_check_banner', async () => {
-    mockBannerFromBackend({
-      banner: {
-        name: BannerName.retry_identity_check_banner,
-        title: 'Retente ubble',
-        text: 'pour débloquer ton crédit',
-      },
-    })
-
-    renderHomeHeader()
-
-    expect(await screen.findByText('Retente ubble')).toBeOnTheScreen()
-    expect(screen.getByText('pour débloquer ton crédit')).toBeOnTheScreen()
-    expect(screen.getByTestId('ArrowAgain')).toBeOnTheScreen()
-  })
-
-  it('should display activation banner with BirthdayCake icon when banner api call return transition_17_18_banner', async () => {
-    mockBannerFromBackend({
-      banner: {
-        name: BannerName.transition_17_18_banner,
-        title: 'Débloque tes 600\u00a0€',
-        text: 'Confirme tes informations',
-      },
-    }),
-      renderHomeHeader()
-
-    expect(await screen.findByText('Débloque tes 600\u00a0€')).toBeOnTheScreen()
-    expect(screen.getByText('Confirme tes informations')).toBeOnTheScreen()
-    expect(screen.getByTestId('BirthdayCake')).toBeOnTheScreen()
-  })
 
   it('should show LocationWidget when isDesktopViewport is false', async () => {
     mockGeolocBannerFromBackend()
@@ -204,15 +139,6 @@ function renderHomeHeader(isDesktopViewport?: boolean) {
       theme: { isDesktopViewport: isDesktopViewport ?? false },
     }
   )
-}
-
-const mockBannerFromBackend = (banner: BannerResponse) => {
-  mockServer.getApi<BannerResponse>('/v1/banner', {
-    responseOptions: {
-      data: banner,
-    },
-    requestOptions: { persist: true },
-  })
 }
 
 function mockGeolocBannerFromBackend() {
